Move ScheduleHomePage out of ScheduleNavigator and into Schedule module
