+++ mobile/src/events/ScheduleNavigator.tsx
@@ -3,22 +3,8 @@ import { Button, SafeAreaView, StyleSheet, View } from "react-native";
 import { createStackNavigator } from "@react-navigation/stack";
 
 import CreateEvent from "./CreateEvent";
-import Schedule from "./Schedule";
+import ScheduleHomePage from "./Schedule";
 
-const ScheduleHomePage: React.FC<{}> = ({ navigation }) => {
-  return (
-    <SafeAreaView style={styles.container}>
-      <Schedule />
-      <Button
-        title="Create Event"
-        onPress={() => {
-          console.log("Create New Event button clicked");
-          navigation.navigate("CreateEvent");
-          return;
-        }}></Button>
-    </SafeAreaView>
-  )
-}
 const ScheduleNavigator: React.FC<{}> = () => {
   const Stack = createStackNavigator();
